Domination !

+++ src/main/java/io/tbbc/cf/property/PropertyDefinition.java
@@ -1,6 +1,8 @@
 package io.tbbc.cf.property;
 
+import java.math.BigDecimal;
+
 public record PropertyDefinition(PropertyName name, String label, String description, boolean reverse,
-                                 CategoryName categoryName, Double minValue, Double maxValue, String unit,
+                                 CategoryName categoryName, BigDecimal minValue, BigDecimal maxValue, String unit,
                                  int decimal, Accessibility accessibility, boolean isFree) {
 }
